explicitly type dates to 64-bit

+++ litcal-darwin/litcal-darwin/ContentView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 
 struct ContentView: View {
-	@State var selectedDateSeconds: Int
+	@State var selectedDateSeconds: Int64
 	@State var isWheelShown = true
 	@Environment(\.colorScheme) private var colorScheme
