+++ litdb/litdb.c
@@ -40,7 +40,7 @@ bool lit_color_from_string(const char *col_str, enum lit_color *out_col) {
 }
 
 bool lit_get_celebration(
-	sqlite3 *db, uint64_t cal_id,
+	sqlite3 *db, int cal_id,
 	int64_t epoch_seconds,
 	struct lit_celebration *out_cel,
 	struct lit_error **out_err
@@ -147,7 +147,7 @@ out:
 }
 
 bool lit_get_min_and_max(
-	sqlite3 *db, uint64_t cal_id,
+	sqlite3 *db, int cal_id,
 	int64_t *out_min, int64_t *out_max,
 	struct lit_error **out_err) {
